Search the value of row[columnId] and replace original value with url
For implementing this, I needed to understand transformFunction
of through2.

https://www.npmjs.com/package/through2#transformfunction

Especially, the following is what I wanted to understand

'To queue a new chunk, call this.push(chunk)—this can be called as many
times as required before the callback() if you have multiple pieces to
send on.

Alternatively, you may use callback(err, chunk) as shorthand for
emitting a single chunk or an error.'

If I use this.push(chunk), the result could not replace the row[colID]
value, even I can see the search is all right.